backend/requirements.txt
# TlaquaNet Backend Dependencies
# ================================
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Production server
gunicorn==23.0.0
